syntastica_js

Function init

#[no_mangle]
pub unsafe extern "C" fn init(
    langs: *const *const c_char,
    langs_len: usize,
)

Initialize the global LanguageSet and [Processor].

This function must be called before any of the others. It accepts a list of languages to load. The function can be called multiple times to re-initialize with a different set of languages.

Safety

The langs pointer must be valid, langs_len must be the length of the langs list, and every pointer in that list must be a pointer to a valid C string.